+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
-## Changed
+### Changed
 - Be more intelligent about whether to install or upgrade
 ### Fixed
 - Stop de-quarantining chromedriver borking