+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.5.0] - 2020-11-11
 ### Added
 - Make what gets installed configurable by a config file
+- Option to config additional commands to run at the beginning and end ot the process
 ### Changed
 - Be more intelligent about whether to install or upgrade
 ### Fixed